@@ -1,51 +1,38 @@
 {
   "name": "example-todo-react",
-  "version": "0.1.0",
   "private": true,
-  "dependencies": {
-    "nanoid": "^3.1.20",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-scripts": "^5.0.1",
-    "web-vitals": "^0.2.4"
-  },
+  "version": "1.0.0",
+  "type": "module",
   "scripts": {
-    "start": "concurrently \"npm run start:css\" \"react-scripts start\"",
-    "start:css": "tailwindcss -o src/tailwind.css --watch",
-    "build": "yarn build:css && react-scripts build",
-    "build:css": "NODE_ENV=production tailwindcss -o src/tailwind.css -m",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview",
+    "test": "vitest",
+    "jsonserver": "json-server --watch db.json --port 3001",
     "prettier": "prettier --write .",
-    "prettier:check": "prettier --check .",
-    "jsonserver": "json-server --watch db.json --port 3001"
+    "prettier:check": "prettier --check ."
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "dependencies": {
+    "nanoid": "^5.0.7",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "concurrently": "^8.2.0",
-    "json-server": "^1.0.0-alpha.23",
-    "prettier": "3.0.1",
-    "tailwindcss": "^3.3.2"
+    "@types/react": "^18.3.11",
+    "@types/react-dom": "^18.3.1",
+    "@vitejs/plugin-react": "^4.3.2",
+    "autoprefixer": "^10.4.20",
+    "eslint": "^9.13.0",
+    "eslint-plugin-react": "^7.37.1",
+    "eslint-plugin-react-hooks": "^5.0.0",
+    "eslint-plugin-react-refresh": "^0.4.13",
+    "json-server": "^1.0.0-beta.3",
+    "postcss": "^8.4.47",
+    "prettier": "^3.3.3",
+    "tailwindcss": "^3.4.14",
+    "typescript": "^5.8.3",
+    "vite": "^5.4.9",
+    "vitest": "^2.1.3"
   }
 }
